Capybara ocumentation and final cleanup.
