+++ .gitignore
@@ -41,6 +41,7 @@ datacollection/get_data.sh
 
 # OSX Garbage
 .DS_Store
+__MACOSX
 
 #Grokcli side effects
 grokcli_summary.csv
@@ -80,9 +81,15 @@ robotium-solo-5.0.1.jar.properties
 .domino*
 *.html
 
+# Jupyter notebooks
 *checkpoint.ipynb
+*.ipynb_checkpoints
+
+# Keras
 *.h5
-projects/sequence_classification/.ipynb_checkpoints
+
+# Datasets extensions
 *_TEST
 *_TRAIN
 *_ALL
+*.log
