+++ projects/capybara/clustering/run.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+set -o errexit
+set -o pipefail
+set -o xtrace
+
+python run_clustering.py
