+++ projects/capybara/clustering/run_clustering.py
@@ -21,18 +21,18 @@
 import copy
 import os
 import shutil
+
 import numpy as np
 from matplotlib import pyplot as plt
 
-from htmresearch.frameworks.classification.utils.traces import loadTraces
-
-from clustering import PerfectClustering, OnlineClusteringV2
-from clustering_interface import Point
-from distances import euclidian_distance
-from utils import (clustering_stats, moving_average,
-                   convert_to_sdrs)
-from plot import (plot_accuracy, plot_cluster_assignments,
-                  plot_inter_sequence_distances)
+from htmresearch.frameworks.capybara.util import clustering_stats
+from htmresearch.frameworks.capybara.sdr import convert_to_sdrs, load_traces
+from htmresearch.frameworks.capybara.util import moving_average
+from htmresearch.frameworks.capybara.unsupervised.clustering import \
+  PerfectClustering, OnlineClustering, Point
+from htmresearch.frameworks.capybara.distance import euclidian_distance
+from htmresearch.frameworks.capybara.unsupervised.plot import \
+  plot_accuracy, plot_cluster_assignments, plot_inter_sequence_distances
 
 
 
@@ -78,7 +78,6 @@ def run(sdrs,
     # Inference.
     new_cluster.add(point)
     confidence, closest_cluster = model.infer(new_cluster.center)
-    #if closest_cluster is None: closest_cluster = new_cluster
     closest_cluster_history.append(closest_cluster)
     if i in cluster_snapshot_indices:
       clusters_snapshots.append([copy.deepcopy(c)
@@ -100,19 +99,19 @@ def run(sdrs,
 
 def main():
   distance_functions = [euclidian_distance]
-  clustering_classes = [PerfectClustering, OnlineClusteringV2]
+  clustering_classes = [PerfectClustering, OnlineClustering]
 
   # Exp params
   moving_average_window = 2  # for all moving averages of the experiment
   ClusteringClass = clustering_classes[1]
   distance_func = distance_functions[0]
-  merge_threshold = 30 # Cutoff distance to merge clusters. 'None' to ignore.
+  merge_threshold = 40  # Cutoff distance to merge clusters. 'None' to ignore.
   start_idx = 0
   end_idx = -1
   input_width = 2048 * 32
   active_cells_weight = 0
   predicted_active_cells_weight = 10
-  max_num_clusters = 3
+  max_num_clusters = 6
   num_cluster_snapshots = 1
   show_plots = True
   distance_matrix_ignore_noise = False  # ignore label 0 if used to label noise.
@@ -125,10 +124,11 @@ def main():
   os.makedirs(plots_output_dir)
 
   # load traces
+
   file_path = os.path.join(os.path.dirname(os.path.abspath(__file__)),
                            os.pardir, 'htm', 'traces',
                            'trace_%s.csv' % exp_name)
-  traces = loadTraces(file_path)
+  traces = load_traces(file_path)
   num_records = len(traces['scalarValue'])
 
   # start and end for the x axis of the graphs
@@ -179,25 +179,12 @@ def main():
                                   ClusteringClass,
                                   merge_threshold,
                                   cluster_snapshot_indices)
-  # cluster_categories = []
-  # for c in closest_cluster_history:
-  #   if c is not None:
-  #     cluster_categories.append(c.label_distribution()[0]['label'])
 
   # plot cluster assignments over time
   for i in range(num_cluster_snapshots):
     clusters = cluster_snapshots[i]
     snapshot_index = cluster_snapshot_indices[i]
     plot_cluster_assignments(plots_output_dir, clusters, snapshot_index)
-
-    # plot inter-cluster distance matrix
-    # plot_id = 'inter-cluster_t=%s' % snapshot_index
-    # plot_inter_sequence_distances(plots_output_dir,
-    #                               plot_id,
-    #                               distance_func,
-    #                               sdrs[:snapshot_index],
-    #                               cluster_categories[:snapshot_index],
-    #                               distance_matrix_ignore_noise)
 
     # plot inter-category distance matrix
     plot_id = 'inter-category_t=%s ' % snapshot_index
