+++ projects/capybara/datasets/README.md
@@ -3,10 +3,10 @@ This folder contains scripts to download, format and analyze the input
 datasets. 
 
 ## Usage
-To download and format the data, run: `./run_all.sh`. The UCR data is password 
+To download and format the data, run: `./run.sh`. The UCR data is password 
 protected, so you'll have to download it manually:
 * Download the zipped data [here](http://www.cs.ucr.edu/~eamonn/time_series_data/UCR_TS_Archive_2015.zip)
-* Password to unzip: `attempttopredict`. Make sure to unzip it in 
+* Password to unzip: `attempttoclassify`. Make sure to unzip it in 
 `nupic.research/projects/capybara/datasets/`.
 
 ## Explore the data
@@ -14,7 +14,7 @@ Visualizations for data exploration are in the the notebook: `explore_datasets.i
 * Start a jupyter notebook: `jupyter notebok`
 * Open and run all the cells in the notebook `explore_datasets.ipynb`
 
-## More details abuot the datasets
+## More details about the datasets
 
 ### UCI
 Follow the `README` in `nupic.research/projects/capybara/datasets/uci_har/`
