+++ projects/capybara/dim_reduction/README.md
@@ -1,0 +1,18 @@
+# Dimensionality reduction
+
+Demo scripts to project SDRs in 2D, in order to make it 
+possible to visualize clusters of SDRs. 
+
+## Pre-requisites
+
+Install pyclustering: 
+```
+    git clone https://github.com/annoviko/pyclustering
+    cd pyclustering
+    pip install . --user
+```
+
+
+## Usage
+* OPTICS demo: `python demo_optics.py`
+* MDS demo:
