--- projects/capybara/htm/run_all.sh
@@ -1,4 +1,0 @@
-python network_runner_sequences.py --config configs/debug.sequences.yml
-python network_runner_sequences.py --config configs/body_acc_x.sequences.yml
-python network_runner_sequences.py --config configs/synthetic_control.sequences.yml
-python network_runner_sequences.py --config configs/test1.sequences.yml
