+++ projects/capybara/requirements.txt
@@ -3,3 +3,5 @@ simplejson
 htmresearch
 nupic
 matplotlib
+sklearn
+PyYaml
