+++ projects/capybara/supervised_baseline/v1_no_sequences/run.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+set -o errexit
+set -o pipefail
+set -o xtrace
+
+python run_baseline.py --config configs/uci.yml
