+++ projects/capybara/supervised_baseline/v2_with_sequences/run.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+set -o errexit
+set -o pipefail
+set -o xtrace
+
+python classify_sequences.py
